+++ RELEASE_NOTES.md
@@ -28,6 +28,9 @@ RELEASE NOTES
 Logtalk compiler and runtime
 ----------------------------
 
+* CHANGED: Calls to unknown dynamic predicates in "this" from a category to
+throw an existence error instead of failing.
+
 * ADDED: Support for `mode/2` directive `zero_or_more_or_error` and
 `one_or_more_or_error` number of proofs.